feat: Update layout, page
