+++ src/app/recommendation/layout.tsx
@@ -1,7 +1,23 @@
-export default function RecommendationLayout({
+"use client";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import { PropsWithChildren } from "react";
+import { FormProvider, useForm } from "react-hook-form";
+import { DevTool } from "@/components/shared/react-hook-form-dev-tool";
+import { recommendationSetupFormSchema } from "@/types/recommendation/setup";
+
+export default function RecommendationSetupLayout({
   children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <div>{children}</div>;
+}: PropsWithChildren) {
+  const form = useForm({
+    mode: "onChange",
+    resolver: zodResolver(recommendationSetupFormSchema),
+  });
+
+  return (
+    <>
+      <FormProvider {...form}>{children}</FormProvider>
+      <DevTool control={form.control} />
+    </>
+  );
 }
